update
git-svn-id: http://tuqianyi3.googlecode.com/svn/tuqianyi3@35 994f2f9b-4fb9-37a0-3060-a1c2e6684766

+++ src/com/tuqianyi/model/Merge.java
@@ -5,8 +5,8 @@ import java.io.Serializable;
 public class Merge implements Serializable{
 	
 	private ImageLabel label;
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	private int z;
 	private int width = 100;
 	private int height = 100;
@@ -19,19 +19,19 @@ public class Merge implements Serializable{
 		return label;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
